requirements.txt
django==4.*
django-cors-headers==3.*
psycopg2==2.*
django-rest-framework
markdown
django-filter
djangorestframework-simplejwt
djangorestframework-camel-case
Pillow
django-cleanup
factory_boy
coverage
django-test-plus
djangorestframework-gis
pandas>=1.4
xlsxwriter
sqlalchemy
django-vectortiles
django-url-filter
owslib
django-sql-utils
mapbox_vector_tile
xarray
drf-spectacular<=0.22
django-postgres-copy
openpyxl
matrixconverters
packaging
cryptography
channels==3.*
channels_redis
routingpy
django_q
matplotlib
pyproj
protobuf==3.20.*